requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --extra=dev --extra=test --output-file=requirements-dev.txt --resolver=backtracking
#
astatine==0.3.2
    # via flake8-encodings
astor==0.8.1
    # via
    #   flake8-qgis
    #   flake8-simplify
astroid==2.15.4
    # via pylint
asttokens==2.2.1
    # via astatine
attrs==23.1.0
    # via
    #   flake8-annotations
    #   flake8-bugbear
bandit[toml]==1.7.5
    # via qgis-plugin-tools (pyproject.toml)
black==23.3.0
    # via qgis-plugin-tools (pyproject.toml)
build==0.10.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via
    #   black
    #   pip-tools
colorama==0.4.6
    # via
    #   bandit
    #   build
    #   click
    #   pylint
    #   pytest
dill==0.3.6
    # via pylint
distlib==0.3.6
    # via virtualenv
domdf-python-tools==3.6.1
    # via
    #   astatine
    #   flake8-encodings
exceptiongroup==1.1.1
    # via pytest
filelock==3.12.0
    # via virtualenv
flake8==6.0.0
    # via
    #   flake8-2020
    #   flake8-annotations
    #   flake8-bugbear
    #   flake8-builtins
    #   flake8-comprehensions
    #   flake8-docstrings
    #   flake8-encodings
    #   flake8-helper
    #   flake8-mutable
    #   flake8-no-pep420
    #   flake8-noqa
    #   flake8-print
    #   flake8-qgis
    #   flake8-simplify
    #   flake8-tidy-imports
    #   flake8-use-fstring
    #   flake8-useless-assert
    #   pep8-naming
    #   qgis-plugin-tools (pyproject.toml)
flake8-2020==1.8.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-annotations==3.0.1
    # via qgis-plugin-tools (pyproject.toml)
flake8-bugbear==23.3.23
    # via qgis-plugin-tools (pyproject.toml)
flake8-builtins==2.1.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-comprehensions==3.12.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-docstrings==1.7.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-encodings[classes]==0.5.0.post1
    # via qgis-plugin-tools (pyproject.toml)
flake8-helper==0.2.1
    # via flake8-encodings
flake8-mutable==1.2.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-no-pep420==2.4.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-noqa==1.3.1
    # via qgis-plugin-tools (pyproject.toml)
flake8-pie==0.16.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-plugin-utils==1.3.2
    # via flake8-pytest-style
flake8-print==5.0.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-pytest-style==1.7.2
    # via qgis-plugin-tools (pyproject.toml)
flake8-qgis==1.0.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-simplify==0.20.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-tidy-imports==4.8.0
    # via qgis-plugin-tools (pyproject.toml)
flake8-use-fstring==1.4
    # via qgis-plugin-tools (pyproject.toml)
flake8-useless-assert==0.4.4
    # via qgis-plugin-tools (pyproject.toml)
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via bandit
identify==2.5.24
    # via pre-commit
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via
    #   pylint
    #   qgis-plugin-tools (pyproject.toml)
jedi==0.18.2
    # via flake8-encodings
lazy-object-proxy==1.9.0
    # via astroid
markdown-it-py==2.2.0
    # via rich
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mdurl==0.1.2
    # via markdown-it-py
mypy==1.2.0
    # via qgis-plugin-tools (pyproject.toml)
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
natsort==8.3.1
    # via domdf-python-tools
nodeenv==1.7.0
    # via pre-commit
packaging==23.1
    # via
    #   black
    #   build
    #   pytest
parso==0.8.3
    # via jedi
pathspec==0.11.1
    # via black
pbr==5.11.1
    # via stevedore
pep8-naming==0.13.3
    # via qgis-plugin-tools (pyproject.toml)
pip-tools==6.13.0
    # via qgis-plugin-tools (pyproject.toml)
platformdirs==3.5.0
    # via
    #   black
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==3.3.1
    # via qgis-plugin-tools (pyproject.toml)
pycodestyle==2.10.0
    # via
    #   flake8
    #   flake8-print
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==3.0.1
    # via flake8
pygments==2.15.1
    # via rich
pylint==2.17.4
    # via qgis-plugin-tools (pyproject.toml)
pyproject-hooks==1.0.0
    # via build
pyqt5-stubs==5.15.6.0
    # via qgis-plugin-tools (pyproject.toml)
pytest==7.3.1
    # via
    #   pytest-mock
    #   pytest-qgis
    #   pytest-qt
    #   pytest-timeout
    #   qgis-plugin-tools (pyproject.toml)
pytest-mock==3.10.0
    # via qgis-plugin-tools (pyproject.toml)
pytest-qgis==1.3.2
    # via qgis-plugin-tools (pyproject.toml)
pytest-qt==4.2.0
    # via qgis-plugin-tools (pyproject.toml)
pytest-timeout==2.1.0
    # via qgis-plugin-tools (pyproject.toml)
pyyaml==6.0
    # via
    #   bandit
    #   pre-commit
rich==13.3.5
    # via bandit
six==1.16.0
    # via asttokens
smmap==5.0.0
    # via gitdb
snowballstemmer==2.2.0
    # via pydocstyle
stevedore==5.0.0
    # via bandit
tomli==2.0.1
    # via
    #   bandit
    #   black
    #   build
    #   mypy
    #   pylint
    #   pyproject-hooks
    #   pytest
tomlkit==0.11.8
    # via pylint
types-requests==2.29.0.0
    # via qgis-plugin-tools (pyproject.toml)
types-urllib3==1.26.25.12
    # via types-requests
typing-extensions==4.5.0
    # via
    #   astroid
    #   black
    #   domdf-python-tools
    #   flake8-noqa
    #   flake8-pie
    #   mypy
    #   pylint
virtualenv==20.23.0
    # via pre-commit
wheel==0.40.0
    # via pip-tools
wrapt==1.15.0
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
